Fixed prefix bug.
When a prefix is empty (e.g. no Text Domain set), all stylesheets are rel attributes were modified.

Added a check for the prefix. If it is empty, we bail.

In addition to that check, I also removed the regular expression checking for the handle, as the handle is provided as an argument in the filter.